logistics_api/requirements.txt
fastapi
pydantic
pydantic[email]
sqlalchemy>=2.0
asyncpg
python-dotenv
uvicorn
passlib[bcrypt]
python-jose[cryptography]
requests
python-multipart
email-validator
httpx
psycopg2-binary